Changed the core code
This is the latest master firmware, still under development by Julian for the next version of the Zero Baseline. It's here to give a status of the changes (triggering, PLL 1 second timer etc.) that have been made. Note the output format is also no longer JSON, which breaks compatability with existing Pi interfaces.